requirements.txt
fastapi
uvicorn
tortoise-orm[asyncpg]
pydantic
pydantic_extra_types
phonenumbers
anyio